+++ gdi/R/gdi.r
@@ -153,6 +153,7 @@ ecomp <- sellipse(vdiam/2, hdiam/2, k)
 #' @param channel Colour channel to which to apply the threshold criterium. Default is 4 (alpha channel of rgba image). Channel setting needs to be adjusted depending on the colour mode of the image used (e.g. there are two channels to choose from in a greyscale image, and 3 in an rgb image).
 #' @param method Method for determining which pixels to count. Default "greater" counts pixels with value greater than threshold (e.g. higher opacity, in the case of an alpha channel). "less" counts pixels with a value less than the threshold. "not" counts all pixels not precisely matching threshold. Any other character string results in only pixels exactly matching the value given as threshold being counted.
 #' @param align Indicate whether the silhouette long axis is aligned horizontally (setting "h", default), or vertically (any other parameter setting).
+#' @param return Setting for what to return, default setting ("diameters") returns a single vector containing the diameters, any other setting returns a data frame containing centers and diameters.
 #' @import jpeg
 #' @import png
 #' @export measuresil
@@ -161,7 +162,7 @@ ecomp <- sellipse(vdiam/2, hdiam/2, k)
 #' fdir <- system.file(package="gdi")
 #' lat <- measuresil(file.path(fdir,"exdata","lat.png"))
 
-measuresil<-function(image_file, threshold=0.5, channel=4, method="greater", align="h"){
+measuresil<-function(image_file, threshold=0.5, channel=4, method="greater", align="h", return="diameters"){
 #load and save image data to variable named img
 if(grepl(".jpg",image_file)==TRUE | grepl(".jpeg",image_file)==TRUE | grepl(".JPG",image_file)==TRUE){
 img <- jpeg::readJPEG(image_file)}#read image if it is jpg
@@ -180,8 +181,20 @@ ncols <- dim(img)[2]
 if(align=="h"){#if horizontally aligned, default
 # Loop through each vertical line of pixels
 depths <- rep(0, ncols)
+centers <- rep(NA, ncols)
+
 for (x in 1:ncols) {
   depth <- 0
+        if(method=="greater"){
+        centers[x]<-nrows-mean(which(img[, x, channel]>threshold))
+        }else if(method=="less"){
+        centers[x]<-nrows-mean(which(img[, x, channel]<threshold))
+        }else if(method=="not"){
+        centers[x]<-nrows-mean(which(signif(img[, x, channel],6)!=signif(threshold,6)))
+        }else{
+        centers[x]<-nrows-mean(which(signif(img[, x, channel],6)==signif(threshold,6)))
+        }
+  
   for (y in 1:nrows) {
     # Get the color of the selected colour channel of the pixel
     color <- img[y, x, channel]
@@ -210,8 +223,21 @@ for (x in 1:ncols) {
 }else{#if vertically aligned
 # Loop through each horizontal line of pixels
 depths <- rep(0, nrows)
+centers <- rep(NA, nrows)
+
 for (y in 1:nrows) {
   depth <- 0
+  
+    if(method=="greater"){
+    centers[y]<-1+mean(which(img[y,,channel]>threshold))
+    }else if(method=="less"){
+    centers[y]<-1+mean(which(img[y,,channel]<threshold))
+    }else if(method=="not"){
+    centers[y]<-1+mean(which(signif(img[y,,channel],6)!=signif(threshold,6)))
+    }else{
+    centers[y]<-1+mean(which(signif(img[y,,channel],6)==signif(threshold,6)))
+    }
+  
   for (x in 1:ncols) {
     # Get the color of the selected colour channel of the pixel
     color <- img[y, x, channel]
@@ -240,15 +266,24 @@ for (y in 1:nrows) {
 
 }
 
+if(return=="diameters"){
 return(depths)
+}else{
+    depthscenters<-data.frame(diameter=depths,center=centers)
+    attr(depthscenters, "x_resolution")<-ifelse(align=="h", ncols,nrows)#total resolution of image along the x axis
+    attr(depthscenters, "y_resolution")<-ifelse(align=="h", nrows,ncols)
+
+return(depthscenters)
+}
+
 }
 
 
 ##Function gdi()
 #'Estimate volume using Graphic Double Integration.
 #'
-#' @param lat Measurements of diameter in lateral view/first of two orthogonal views to be used with the gdi. Can be either a numeric vector, or a text file to be scanned. Defaults to "lat.txt".
-#' @param dors Measurements of diameter in dorsal view/second of two orthogonal views to be used with the gdi. Can be either a numeric vector, or a text file to be scanned. Must be the same length as lat. Defaults to "dors.txt".
+#' @param lat Measurements of diameter in lateral view/first of two orthogonal views to be used with the gdi. Can be either a numeric vector, a data.frame (output of measuresil(...,return="all") with a collumn named "diameter", or a text file with diameter measurements to be scanned.
+#' @param dors Measurements of diameter in dorsal view/second of two orthogonal views to be used with the gdi. Can be either a numeric vector, a data.frame (output of measuresil(...,return="all") with a collumn named "diameter", or a text file with diameter measurements to be scanned. Must be the same length as lat.
 #' @param indices Optional indices specifying a subset of the silhouette measurement vectors to be analyzed. Useful if separate segment calculations are desired.
 #' @param scale Scale of the data in terms of how many units of the input data are in one side of the desired unit of output volume. Defaults to 10.
 #' @param sliceL Length of individual segments to be used in the GDI. Defaults to 1/scale.
@@ -266,38 +301,66 @@ return(depths)
 #' gdi(lat=lateral, dors=lateral/2, scale=10, method="smooth", k=2.3)
 
 
-gdi<-function(lat, dors, indices=c(1:length(lat)), scale=10, sliceL=1/scale, method="raw", k=2.0, corr=1, smooth.ends=FALSE, return="total"){
+gdi<-function(lat, dors, indices=NULL, scale=10, sliceL=1/scale, method="raw", k=2.0, corr=1, smooth.ends=FALSE, return="total"){
+lat_<-lat
+dors_<-dors
 
-if(is.character(lat)){scan(lat)->lat}
+if(is.character(lat)){scan(lat)->lat_}
+if(is.character(dors)){scan(dors)->dors_}#scan files, if inputs are filenames
 
-if(is.character(dors)){scan(dors)->dors}
 
-sil<-data.frame(lat/scale,dors/scale)#scale-adjust and save in dataframe
+if(is.data.frame(lat)){
+        if(length(lat$diameter)!=nrow(lat)){
+        stop("lat is a data.frame, but collumn \"diameter\" not found!")}
+    lat_<-lat$diameter
+}
 
-sil<-sil[indices,]
+if(is.data.frame(dors)){
+        if(length(dors$diameter)!=nrow(dors)){
+        stop("dors is a data.frame, but collumn \"diameter\" not found!")}
+    dors_<-dors$diameter
+}
+
+sil<-data.frame(ydiam_raw=lat_,zdiam_raw=dors_,ydiam_scaled=lat_/scale,zdiam_scaled=dors_/scale, slice_length=sliceL)#scale-adjust and save in dataframe
+
+
+if(is.data.frame(lat)){
+sil$y_center<-lat$center
+}
+
+if(is.data.frame(dors)){
+sil$z_center<-dors$center
+}
+
+if(!is.null(indices)){
+sil<-sil[indices,]}
 
 #estimate cross-sections as (super)ellipses
-sil$A<-sellipse(sil$lat/2, sil$dors/2, k)
+sil$A<-sellipse(sil$ydiam_scaled/2, sil$zdiam_scaled/2, k)
 
 if(method=="raw"){
     sil$V<-sil$A*sliceL#raw formula, approximating each segment as a cylinder
 }else{
-    for(i in 1:(length(lat)-1)){
+    for(i in 1:(nrow(sil)-1)){
     sil$V[i]<-(sil$A[i]+sil$A[i+1]+sqrt(sil$A[i]*sil$A[i+1]))/3*ifelse(length(sliceL)>1,sliceL[1],sliceL)
     #approximating segments as frusta, with A[i] and A[i+1] as base areas
     }
     if(smooth.ends==TRUE){
     sil$V[1]<-ifelse(length(sliceL)>1,sliceL[1],sliceL)*sil$A[1]/3#conical first segment (tapers to 0)
-    sil$V[length(lat)]<-ifelse(length(sliceL)>1,sliceL[length(lat)],sliceL)*sil$A[length(lat)]/3#conical last segment (tapers to 0)
+    sil$V[nrow(sil)]<-ifelse(length(sliceL)>1,sliceL[nrow(sil)],sliceL)*sil$A[nrow(sil)]/3#conical last segment (tapers to 0)
     }else{
-    sil$V[length(lat)]<-ifelse(length(sliceL)>1,sliceL[length(lat)],sliceL)*sil$A[length(lat)]
+    sil$V[nrow(sil)]<-ifelse(length(sliceL)>1,sliceL[nrow(sil)],sliceL)*sil$A[nrow(sil)]
     }
 }
 
 sil$V <- sil$V*corr
+attr(sil, "x_resolution")<-attr(lat, "x_resolution")
+attr(sil, "y_resolution")<-attr(lat, "y_resolution")
+attr(sil, "z_resolution")<-attr(dors, "y_resolution")
+
 
 sum(sil$V)->res#sum up segments
-names(res) <- paste("x_dim",sum(lat!=0)/scale, "units", sep="_")
+names(res) <- paste("x_dim",sum(sil$ydiam_raw!=0)/scale, "units", sep="_")
 
 if(return=="total"){
 return(res)
@@ -432,6 +495,132 @@ sorted <- cbind(sorted, sorted/total)
 rownames(sorted)<-names
 colnames(sorted) <- c("count","proportion")
 return(sorted)
+}
+
+}
+
+
+##Function hCOM()
+#'Finds the horizontal (x axis, i.e. the axis vertical to the cross-sections) position of the center of mass (COM) of the volume. Experimental; only valid for "raw" gdi results with segment volumes approximated as elliptical prisms, or for manually supplied segment COMs. COM is calculated as a weighted mean of all segment COMs, with the segment mass as the weighting factor.
+#'
+#' @param x Either a data frame that is the output of gdi(..., return="all"), or a numeric vector of horizontal segment COM positions.
+#' @param volumes An optional separate vector of volumes, required if x is not a data.frame containing volumes.
+#' @param align alignment of the silhouette, if "h" (default) the silhouette is assumed to be horizontally aligned, if any other value (e.g. "v") then the silhouette is assumed to be vertically aligned.
+#' @param subtract An optional separate vector of volumes, with length equal to the length or nrow() of x, to be subtracted from the volumes for the COM calculation.
+#' @param densities An optional vector of segment densities, with length equal to the length or nrow() of x, to be multiplied with the volumes for the COM calculation. If both subtract and densities are supplied, the density is applied only to the "residual" volume that is left after subtraction.
+#' @param scale Optional scale value (number of pixels to chosen unit of measurement)
+#' @return An object of class numeric() containing the x coordinate of the center of mass of the shape, in pixels (or chosen units, if manually calculated)
+#' @export hCOM
+#' @importFrom stats weighted.mean
+#' @examples
+#' fdir <- system.file(package="gdi")
+#' measuresil(file.path(fdir,"exdata","lat.png"), return="all")->lat_
+#' measuresil(file.path(fdir,"exdata","dors.png"), return="all")->dors_
+#' gdi(lat_, dors_, return="all")->gdiresults
+#' hCOM(gdiresults)
+
+
+
+hCOM<-function(x, volumes=NULL, align="h", subtract=NULL, densities=NULL, scale=1){
+x_center<-x
+if(is.data.frame(x)){
+volumes<-x$V#look for collumn "V" containing segment volumes
+x_center<-c(1:nrow(x))-0.5#save horizontal COM positions based on segment numbers
+}
+masses<-volumes
+if(!is.null(subtract)){masses<-volumes-subtract}#subtract airspace volumes
+if(!is.null(densities)){masses<-masses*densities}#and/or multiply by segment densities given as vector
+
+if(align!="h"){
+print("converted from top-down measurement")
+    xmax<-attributes(x)$x_resolution
+    return((xmax-weighted.mean(x_center, w=masses, na.rm=TRUE))/scale)
+    #}
+}else{
+return(weighted.mean(x_center,w=masses, na.rm=TRUE)/scale)
+}
+}
+
+
+
+##Function vCOM()
+#'Finds the vertical (y axis, i.e. the axis parallel to the cross-section diameter) position of the center of mass (COM) of the volume. Experimental; only valid for "raw" gdi results with segment volumes approximated as elliptical prisms, or for manually supplied segment COMs. COM is calculated as a weighted mean of all segment COMs, with the segment mass as the weighting factor. Estimates have lower accuracy compared to hCOM, because cross-sectional geometry and variation in density throughout the cross-section is not taken into account.
+#'
+#' @param y A data.frame that is the output of gdi(..., return="all"), or a numeric vector containing vertical COM positions for segments
+#' @param volumes An optional separate vector or data.frame (output of gdi(...,return="all") or vector of volumes.
+#' @param subtract An optional separate vector of volumes, with length equal to the length or nrow() of x, to be subtracted from the volumes for the COM calculation.
+#' @param densities An optional vector of segment densities, with length equal to the length or nrow() of x, to be multiplied with the volumes for the COM calculation. If both subtract and densities are supplied, the density is applied only to the "residual" volume that is left after subtraction.
+#' @param scale Optional scale value (number of pixels to chosen unit of measurement)
+#' @param from_top Whether the output coordinate should be measured from the top of the image (standard for image processing software), if TRUE, or from the bottom (standard for plotting in R (if FALSE, default). If TRUE, an attribute to y, containing the vertical dimension relative to which the measurement should be taken is required.
+#' @return An object of class numeric() containing the y coordinate of the center of mass of the shape, in pixels (or chosen units, if manually calculated)
+#' @export vCOM
+#' @importFrom stats weighted.mean
+#' @examples
+#' fdir <- system.file(package="gdi")
+#' measuresil(file.path(fdir,"exdata","lat.png"), return="all")->lat_
+#' measuresil(file.path(fdir,"exdata","dors.png"), return="all")->dors_
+#' gdi(lat_, dors_, return="all")->gdiresults
+#' vCOM(gdiresults)
+
+
+vCOM<-function(y,volumes=NULL,subtract=NULL, densities=NULL, scale=1, from_top=FALSE){
+y_center<-y
+    if(is.data.frame(y)){
+    volumes<-y$V#look for collumn "center" containing vertical segment COM positions
+    y_center<-y$y_center#look for collumn "center" containing vertical segment COM positions
+    }
+
+if(!is.null(volumes) & is.data.frame(volumes)){
+volumes<-volumes$V#look for collumn "V" containing segment volumes
+}
+
+
+masses<-volumes
+if(!is.null(subtract)){masses<-volumes-subtract}#subtract airspace volumes
+if(!is.null(densities)){masses<-volumes*densities}#and/or multiply by segment densities given as vector
+
+if(from_top==TRUE){
+ ymax<-attributes(y)$y_resolution
+ return((ymax-weighted.mean(y_center, w=masses, na.rm=TRUE))/scale)
+    #}
+}else{
+return(weighted.mean(y_center, w=masses, na.rm=TRUE)/scale)
+}
+}
+
+
+
+##Function plot.sil()
+#' Plots a silhouette read by measuresil()
+#'
+#' @param sil A data frame that is the output of measuresil(..., return="all"), containing the center and the diameter of the silhouette at each value for x.
+#' @param flip Whether to flip axes (useful if measuresil() was performed using align="v", defaults to FALSE.
+#' @param add Whether to add to an existing plot
+#' @param ... Other parameters to pass on to plot() or lines()
+#' @return A plotted silhuette
+#' @export plot.sil
+#' @examples
+#' fdir <- system.file(package="gdi")
+#' measuresil(file.path(fdir,"exdata","lat.png"), return="all")->lat_
+#' plot.sil(lat_)
+
+plot.sil<-function(sil, flip=FALSE, add=FALSE, ...){
+
+if(flip==FALSE){
+x<-c(1:nrow(sil),rev(1:nrow(sil)))
+y<-c(sil$center+sil$diameter/2, rev(sil$center-sil$diameter/2))
+}else{
+y<--c(1:nrow(sil),rev(1:nrow(sil)))+nrow(sil)
+x<-c(sil$center+sil$diameter/2, rev(sil$center-sil$diameter/2))
+}
+
+if(add==TRUE){
+lines(y~x, ...)
+
+}else{
+
+plot(y~x, type="l", xlab="x",ylab="y",...)
+
 }
 
 }